ChibiOS RT: PC High Level Driver
On some uControllers not all the pins have a dedicated irq line.
Some are grouped in bundles - ports - and those which belong to the same
port share the same irq line.

This is the high level abstract driver for ChibiOS.
It relies on a counterpart (low level driver) implementing the details
specific to the chosen uC.

Signed-off-by: Igor Stoppa <igor.stoppa@gmail.com>